requirements.txt
alembic==1.7.4
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
enum34==1.1.10
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.7.0
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
google-search-results==2.4.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==4.1.0
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
urllib3==1.26.7
Werkzeug==1.0.1
zipp==3.6.0
